make some code more pythonic
use references to objects instead of indices whenever possible
iterate over elements instead of indices whenever possible

use lazy computation for the linear parts of the processing:
yield makes a function return a generator (or iterator) and when
that generator is asked for the next element (by a for loop)
it resumes the function yielding elements until the next yield.
In our case this means the merged elements and sql statement lists
are no longer kept as a whole in memory and output is generated
as a new merged element becomes available.